+++ gatsby-config.js
@@ -9,11 +9,13 @@ module.exports = {
     author: `@newrelic`,
     repository: 'https://github.com/newrelic/opensource-website',
     siteUrl: 'https://opensource.newrelic.com',
+    branch: 'develop',
   },
   plugins: [
     {
       resolve: '@newrelic/gatsby-theme-newrelic',
       options: {
+        gaTrackingId: 'UA-3047412-33',
         layout: {
           // 1180px (width) + 28px (padding) * 2 (left/right side)
           maxWidth: '1236px',
@@ -146,28 +148,6 @@ module.exports = {
         ],
         // remarkPlugins: [require(`gatsby-remark-copy-linked-files`)],
         extensions: [`.mdx`, `.md`],
-      },
-    },
-    {
-      resolve: 'gatsby-plugin-edit-content-links',
-    },
-    {
-      resolve: 'gatsby-plugin-gdpr-tracking',
-      options: {
-        // logging to the console, if debug is true
-        debug: false,
-        googleAnalytics: {
-          // The property ID; the tracking code won't be generated without it.
-          trackingId: 'UA-3047412-33',
-          // Defines it google analytics should be started with out the cookie consent
-          autoStart: false, // <--- default
-          // Setting this parameter is optional
-          anonymize: true, // <--- default
-          // Name of the cookie, that enables the tracking if it is true
-          controlCookieName: 'newrelic-gdpr-consent', // <--- default
-          cookieFlags: 'secure;samesite=none', // <--- default
-        },
-        environments: ['production', 'development'],
       },
     },
   ],